Split superblock.rs into superblock/ subdirectory
Split crates/simulation/src/superblock.rs (637 lines) into focused modules
under superblock/ for better maintainability:
- constants.rs (17 lines): policy tuning constants
- types.rs (94 lines): Superblock struct and SuperblockCell enum
- state.rs (161 lines): SuperblockState resource and Saveable impl
- systems.rs (43 lines): ECS system and plugin registration
- tests.rs (304 lines): all unit tests
- mod.rs (37 lines): thin facade with re-exports
